refactor: optimize DbSetMockHelper for improved unit testing
- Add null-check to the data parameter in MockDbSet<T> for safer instantiation.
- Streamline the setup of IQueryable<T> and IAsyncEnumerable<T> interfaces.
- Introduce an async Add method that returns a mocked EntityEntry<T>.
- Simplify CRUD operation setups using lambda expressions.
- Remove redundant comments to enhance code readability.
- Update the using directive in RefreshTokenServiceTests for NojectServer.Tests.MockHelpers.
